+++ q_2.2/client.c
@@ -5,11 +5,16 @@
 #include <stdlib.h>
 
 int main(int argc , char* argv []){
-    int pid_server = atoi(argv[1]);
-    int sig = atoi(argv[2]);
-    int num_send = atoi(argv[3]);
-    for(int i=0; i<num_send; i++){
-        kill(pid_server,sig);
+    if(argc==4 && (atoi(argv[2])==2 || atoi(argv[2])==10))  {
+        pid_t pid_server = atoi(argv[1]);
+        int sig = atoi(argv[2]);
+        int num_of_send = atoi(argv[3]);
+        for (int i = 0; i < num_of_send; i++) {
+            kill(pid_server, sig);
+        }
+    }
+    else{
+        printf("Please enter valid arguments\n");
     }
     return 0;
 }
